requirements.txt
Fabric==1.8.3
simplejson==2.6.1
python-etcd==0.3.0